+++ business/Join/DivElemento.cs
@@ -5,8 +5,8 @@ namespace business.Join
 {
     public class DivElemento
     {
-        public int? ElementoId { get; set; }
-        public int? DivId { get; set; }
+        public ulong? ElementoId { get; set; }
+        public ulong? DivId { get; set; }
         public virtual Div Div { get; set; }
         public virtual Elemento Elemento { get; set; }
     }
